Fix for IE8's lack of support for Object.create
depot.js uses various parts of the ES5 standard, like Array.indexOf and Object.create. But while polyfills exist that can handle the use of Array.indexOf, none (to my knowledge) exist that will handle Object.create with a second parameter. This branch aims to fix that so IE8 can be shimmed to work with depot.js.